refactor: moved tests according to increase cohesion
